requirements.txt
pyperclip
Pillow
googletrans==4.0.0-rc1
speechrecognition
pyaudio
google-generativeai
